requirements.txt
eventlet~=0.36.1
opencv-python~=4.4.0.42
numpy~=1.19.5
Flask~=2.2.5
Pillow~=9.5.0
tensorflow~=2.6.0
pandas~=1.3.5
scikit-learn~=0.22
keras~=2.6.0
requests~=2.31.0
setuptools~=68.0.0